requirements.txt
beautifulsoup4==4.12.2
cachetools==5.3.0
fastapi==0.108.0
flask
flask-cors
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.3
plotly==5.18.0
pydantic
python-dotenv==1.0.0
pytz
requests==2.31.0
starlette
yfinance==0.2.36
uvicorn==0.25.0
# Database
sqlalchemy==2.0.25
alembic==1.12.1
psycopg2-binary==2.9.9
# JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 